チェックサム計算をCRC32に変更し、フロントエンドとバックエンドの整合性を向上。関連するデータ型をuint64からuint32に変更し、アップロード処理の初期化を強化。ログ出力を改善し、ユーザー体験を向上させるためのデバッグ情報を追加。

+++ backend/presentation/ws/files.go
@@ -10,7 +10,7 @@ import (
 )
 
 type UploadFileChunkData struct {
-	Checksum uint64
+	Checksum uint32
 	Chunk    []byte
 }
